requirements.txt
aioconsole==0.6.2
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
comm==0.1.4
constantly==15.1.0
cryptography==41.0.3
cssselect==1.2.0
debugpy==1.6.7.post1
decorator==5.1.1
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
executing==1.2.0
filelock==3.12.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.14.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.0
jmespath==1.0.1
jupyter_client==8.3.0
jupyter_core==5.3.1
lxml==4.9.3
Markdown==3.4.4
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.25.1
packaging==23.1
pandas==2.0.3
parsel==1.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
pluggy==1.2.0
prompt-toolkit==3.0.39
Protego==0.3.0
psutil==5.9.5
psycopg==3.1.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
pytest==7.4.0
python-dateutil==2.8.2
python-scrapyd-api==2.1.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.10.0
scrapy-djangoitem==1.1.1
scrapy-rotating-proxies==0.6.2
scrapyd==1.4.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tldextract==3.4.4
tornado==6.3.3
traitlets==5.9.0
Twisted==22.10.0
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
w3lib==2.1.2
wcwidth==0.2.6
websockets==11.0.3
zope.interface==6.0
